# Use standard C function to copy characters