--- a/ext/oj/dump.h
+++ b/ext/oj/dump.h
@@ -82,8 +82,9 @@ inline static bool dump_ignore(Options opts, VALUE obj) {
 }
 
 inline static void dump_ulong(unsigned long num, Out out) {
-    char  buf[32];
-    char *b = buf + sizeof(buf) - 1;
+    char   buf[32];
+    char  *b   = buf + sizeof(buf) - 1;
+    size_t cnt = 0;
 
     *b-- = '\0';
     if (0 < num) {
@@ -94,9 +95,9 @@ inline static void dump_ulong(unsigned long num, Out out) {
     } else {
         *b = '0';
     }
-    for (; '\0' != *b; b++) {
-        *out->cur++ = *b;
-    }
+    cnt = sizeof(buf) - (b - buf) - 1;
+    memcpy(out->cur, b, cnt);
+    out->cur += cnt;
     *out->cur = '\0';
 }