… ext/oj/dump_object.c
… ext/oj/dump_strict.c
… ext/oj/rails.c