It has a slightly better performance to copy characters by using standard C function

−       | before   | after    | result
--       | --       | --       | --
macOS    | 363.396k | 377.166k | 1.038x
Linux    | 347.901k | 354.864k | 1.020x

### Environment
- macOS
  - macOS 12.3 beta
  - Apple M1 Max
  - Apple clang version 13.1.6 (clang-1316.0.20.6)
  - Ruby 3.1.1
- Linux
  - Kubuntu 21.10
  - AMD Ryzen 7 5700G
  - gcc version 11.2.0
  - Ruby 3.1.1

### macOS
#### Before
```
Warming up --------------------------------------
            Oj.dump     36.222k i/100ms
Calculating -------------------------------------
            Oj.dump     363.396k (± 0.4%) i/s -      7.281M in  20.035345s
```

### After
```
Warming up --------------------------------------
            Oj.dump     37.656k i/100ms
Calculating -------------------------------------
            Oj.dump     377.166k (± 0.4%) i/s -      7.569M in  20.068008s
```

### Linux
#### Before
```
Warming up --------------------------------------
            Oj.dump     34.784k i/100ms
Calculating -------------------------------------
            Oj.dump     347.901k (± 0.4%) i/s -      6.992M in  20.096716s
```

#### After
```
Warming up --------------------------------------
            Oj.dump     35.345k i/100ms
Calculating -------------------------------------
            Oj.dump     354.864k (± 0.3%) i/s -      7.104M in  20.020172s
```

### Test code
```ruby
require 'benchmark/ips'
require 'oj'

class Foo
  attr_accessor :bar

  def initialize
    @foo = 'foo'
    @bar = 'bar'
  end
end

class Bar < Struct.new(:x, :y)
end

ary = [2 ** 60, 2 ** 61, 2 ** 62, 2 ** 63]
ary << ary

hash = {
  name: 'hogehoge',
  age: 42,
}
hash[:circular] = hash

data = {
  klass: Foo,
  object: Foo.new,
  array: ary,
  range: (1..10),
  hash: hash,
  data: Time.new,
  struct: Bar.new(1024, 2048),
  string: "a" * 100,
  symbol: ("a" * 100).to_sym,
  regex: /foo|bar|baz/,
  rational: 42.to_r,
  complex: Complex('2/3+3/4i'),
}

Benchmark.ips do |x|
  x.time = 20

  x.report('Oj.dump ') { Oj.dump(data, mode: :object, circular: true) }
end
```